Rename oav async file
