+++ tests/devices/unit_tests/oav/test_oav_async.py
@@ -1,7 +1,7 @@
 import pytest
 from ophyd_async.core import DeviceCollector, set_mock_value
 
-from dodal.devices.oav.oav_async import OAV, ZoomController
+from dodal.devices.oav.oav_detector import OAV, ZoomController
 from dodal.devices.oav.oav_parameters import OAVConfig
 
 DISPLAY_CONFIGURATION = "tests/devices/unit_tests/test_display.configuration"
